Add a base document-index to allow indexing non-content items in AzureAI search
