+++ src/OrchardCore/OrchardCore.Search.AzureAI.Core/Models/SearchIndexDefinition.cs
@@ -1,4 +1,4 @@
-using static OrchardCore.Indexing.DocumentIndex;
+using static OrchardCore.Indexing.DocumentIndexBase;
 
 namespace OrchardCore.Search.AzureAI.Models;
