+++ src/OrchardCore/OrchardCore.Search.AzureAI.Core/Services/AzureAISearchIndexDocumentManager.cs
@@ -6,7 +6,7 @@ using OrchardCore.Contents.Indexing;
 using OrchardCore.Indexing;
 using OrchardCore.Modules;
 using OrchardCore.Search.AzureAI.Models;
-using static OrchardCore.Indexing.DocumentIndex;
+using static OrchardCore.Indexing.DocumentIndexBase;
 
 namespace OrchardCore.Search.AzureAI.Services;
 
@@ -130,7 +130,7 @@ public class AzureAIIndexDocumentManager
         await DeleteDocumentsAsync(indexName, contentItemIds);
     }
 
-    public async Task<bool> MergeOrUploadDocumentsAsync(string indexName, IList<DocumentIndex> indexDocuments, AzureAISearchIndexSettings indexSettings)
+    public async Task<bool> MergeOrUploadDocumentsAsync(string indexName, IList<DocumentIndexBase> indexDocuments, AzureAISearchIndexSettings indexSettings)
     {
         ArgumentException.ThrowIfNullOrEmpty(indexName);
         ArgumentNullException.ThrowIfNull(indexDocuments);
@@ -244,7 +244,7 @@ public class AzureAIIndexDocumentManager
         indexMappings.Add(indexMap);
     }
 
-    private static IEnumerable<SearchDocument> CreateSearchDocuments(IEnumerable<DocumentIndex> indexDocuments, Dictionary<string, IEnumerable<AzureAISearchIndexMap>> mappings)
+    private static IEnumerable<SearchDocument> CreateSearchDocuments(IEnumerable<DocumentIndexBase> indexDocuments, Dictionary<string, IEnumerable<AzureAISearchIndexMap>> mappings)
     {
         foreach (var indexDocument in indexDocuments)
         {
@@ -252,13 +252,15 @@ public class AzureAIIndexDocumentManager
         }
     }
 
-    private static SearchDocument CreateSearchDocument(DocumentIndex documentIndex, Dictionary<string, IEnumerable<AzureAISearchIndexMap>> mappingDictionary)
+    private static SearchDocument CreateSearchDocument(DocumentIndexBase documentIndex, Dictionary<string, IEnumerable<AzureAISearchIndexMap>> mappingDictionary)
     {
-        var doc = new SearchDocument()
+        var doc = new SearchDocument();
+
+        if (documentIndex is DocumentIndex index)
         {
-            { IndexingConstants.ContentItemIdKey, documentIndex.ContentItemId },
-            { IndexingConstants.ContentItemVersionIdKey, documentIndex.ContentItemVersionId },
-        };
+            doc.Add(IndexingConstants.ContentItemIdKey, index.ContentItemId);
+            doc.Add(IndexingConstants.ContentItemVersionIdKey, index.ContentItemVersionId);
+        }
 
         foreach (var entry in documentIndex.Entries)
         {
@@ -276,14 +278,14 @@ public class AzureAIIndexDocumentManager
 
             switch (entry.Type)
             {
-                case Types.Boolean:
+                case DocumentIndexBase.Types.Boolean:
                     if (entry.Value is bool boolValue)
                     {
                         doc.TryAdd(map.AzureFieldKey, boolValue);
                     }
                     break;
 
-                case Types.DateTime:
+                case DocumentIndexBase.Types.DateTime:
 
                     if (entry.Value is DateTimeOffset offsetValue)
                     {
@@ -296,7 +298,7 @@ public class AzureAIIndexDocumentManager
 
                     break;
 
-                case Types.Integer:
+                case DocumentIndexBase.Types.Integer:
                     if (entry.Value != null && long.TryParse(entry.Value.ToString(), out var value))
                     {
                         doc.TryAdd(map.AzureFieldKey, value);
@@ -304,21 +306,21 @@ public class AzureAIIndexDocumentManager
 
                     break;
 
-                case Types.Number:
+                case DocumentIndexBase.Types.Number:
                     if (entry.Value != null)
                     {
                         doc.TryAdd(map.AzureFieldKey, Convert.ToDouble(entry.Value));
                     }
                     break;
 
-                case Types.GeoPoint:
+                case DocumentIndexBase.Types.GeoPoint:
                     if (entry.Value != null)
                     {
                         doc.TryAdd(map.AzureFieldKey, entry.Value);
                     }
                     break;
 
-                case Types.Text:
+                case DocumentIndexBase.Types.Text:
                     if (entry.Value != null)
                     {
                         var stringValue = Convert.ToString(entry.Value);
